pkg/cdi: implement automatic refresh.
Implement automatic refresh and enable it by default. The new
WithAutoRefresh(bool) option can be used to explicitly change
the auto-refreshing behavior of the cache.

When auto-refresh is enabled, CDI Spec directories are monitored
for changes. A cache refresh is internally triggered whenever a
relevant change to a cache file or directory is detected.

Signed-off-by: Krisztian Litkey <krisztian.litkey@intel.com>