pkg/cdi: add Spec name generation functions.
Add functions for generating ordinary or transient
Spec file names. The generated names can be passed
to WriteSpec() to create the actual Spec files and
to RemoveSpec() to remove them.

Also update the package introduction in the golang
docs describing how dynamic Spec files should be
generated using the API.

Signed-off-by: Krisztian Litkey <krisztian.litkey@intel.com>