remove usage of github.com/hashicorp/go-multierror
The internal replacement is a simpler implementation which only supports enough
functionality for the places where github.com/hashicorp/go-multierror was used
before.

Not depending on github.com/hashicorp/go-multierror and thus
github.com/hashicorp/errwrap makes it easier to use CDI in Kubernetes because
we don't need to add those other dependencies.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>