pkg/cdi: remove github.com/opencontainers/runc/libcontainer dependency
commit ee12044b3895afdff3841fee894e3edf01d7fc26 brought in this dependency
on runc's libcontainer/devices package.

The runc maintainers consider "libcontainer" to be unstable, and not to
be used by external consumers. The package (due to it being part of a
large module) also brings in many transitive dependencies.

Given that the DeviceFromPath is only a small wrapper for some golang.org/x/sys
calls, and the main purpose of the utility (giving information about
"cgroup_permissions (which cannot be easily queried)") was not used, we
can replace it with a more customised implemenation that's targetted at
the information that's needed here.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>