pkg/cdi: implement default cache.
The default cache is always available and implicitly created the
first time it is referenced. It is created with the default cache
options but can then be reconfigured explicitly if necessary.

Aso add package level variants of the most commonly used cache
functions, Refresh(), InjectDevices, and GetErrors(). These all
use the default cache.

Signed-off-by: Krisztian Litkey <krisztian.litkey@intel.com>